refactor system.py
